modules/extra-LogForensis/requirements.txt
aws-cdk-lib==2.90.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.90.0a0

